% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_group_calendar.R
\name{compute_group_calendar}
\alias{compute_group_calendar}
\title{Title}
\usage{
compute_group_calendar(data, scales)
}
\arguments{
\item{scales}{}
}
\value{

}
\description{
Title
}
\examples{
return_dates_year(1999) \%>\%
head() \%>\%
compute_group_calendar()

}
