% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_week.R
\name{geom_text_week}
\alias{geom_text_week}
\title{Title}
\usage{
geom_text_week(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(tidyverse)
library(lubridate)

data.frame(date = as.Date("2020-01-01") + hours(0:2000)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_text_week() +
  facet_wrap(~fullweek(date)) +
  scale_y_reverse(limits = c(22, 7))

data.frame(date = as.Date("2020-01-01") + hours(0:2000)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_text_week() +
  facet_wrap(~fullweek(date)) +
  scale_y_reverse(limits = c(22, 7))


}
