% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_weekly.R
\name{geom_text_weekly}
\alias{geom_text_weekly}
\title{Title}
\usage{
geom_text_weekly(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(tidyverse)
library(lubridate)

data.frame(date = as.Date("2020-01-01") + days(0:365)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_text_calendar() +
  facet_wrap(~month(date, label = TRUE, abbr = TRUE)) +
  scale_y_reverse()

data.frame(date = as.Date("2020-01-01") + days(0:800)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_text_calendar() +
  facet_grid(year(date) ~ month(date, label = TRUE, abbr = TRUE))

  data.frame(date = as.Date("2020-01-01") + days(0:800)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_text_calendar() +
  scale_x_discrete() +
  facet_grid(lubridate::year(date) ~ lubridate::month(date, label = TRUE, abbr = TRUE))

}
