% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tile_calendar.R
\name{geom_tile_calendar}
\alias{geom_tile_calendar}
\title{Title}
\usage{
geom_tile_calendar(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(lubridate)
library(tidyverse)

data.frame(date = as.Date("2020-01-01") + days(0:365)) \%>\%
  ggplot() +
  aes(date = date) +
  aes(color = date) +
  geom_tile_calendar() +
  facet_wrap(~month(date, label = T, abbr = T))

data.frame(date = as.Date("2020-01-01") + days(0:400)) \%>\%
  ggplot() +
  aes(date = date) +
  geom_tile_calendar() +
  facet_grid(year(date) ~ month(date, label = T, abbr = T))
}
