% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggweekly.R
\name{ggweekly}
\alias{ggweekly}
\title{Title}
\usage{
ggweekly(
  dates_df = return_hours_week(),
  day_labels = c("S", "M", "T", "W", "T", "F", "S")
)
}
\arguments{
\item{dates_df}{}
}
\value{

}
\description{
Title
}
\examples{
library(lubridate)
library(dplyr)
library(ggplot2)
library(magrittr)

ggweekly() +
geom_text_weekly()

ggweekly() +
geom_text_weekly(color = "grey35") +
labs(title = "When to do #TidyTuesday in 2022") +
geom_text_weekly(label = "X",
                  data = data.frame(date = seq(as.Date("2022/01/01"),
                  as.Date("2022/04/18"), "days")))

}
