% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_dates_year.R
\name{return_dates_interval}
\alias{return_dates_interval}
\title{Title}
\usage{
return_dates_interval(start_date, end_date)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
Title
}
\examples{
return_dates_interval("2020-01-01", "2020-01-05")
}
