% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_dates_year.R
\name{return_dates_month}
\alias{return_dates_month}
\title{Title}
\usage{
return_dates_month(month)
}
\arguments{
\item{month}{}
}
\value{

}
\description{
Title
}
\examples{
return_dates_month("1999-02")
return_dates_month("1999-04")
}
