% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_stat_calendar.R
\name{stat_calendar}
\alias{stat_calendar}
\title{Title}
\usage{
stat_calendar(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{geom}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\description{
Title
}
