% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_count_calendar.R
\name{geom_count_calendar}
\alias{geom_count_calendar}
\title{Title}
\usage{
geom_count_calendar(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(lubridate)
library(tidyverse)

data.frame(date = as.Date("2020-01-01") + days(0:365) ) \%>\%
  sample_n(500, replace = TRUE) \%>\%
  ggplot() +
  aes(date = date) +
  geom_count_calendar() +
  scale_color_discrete() +
  facet_wrap(~month(date, label = TRUE, abbr = TRUE))
}
