% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{ggcalendar}
\alias{ggcalendar}
\title{Title}
\usage{
ggcalendar(
  dates_df = df_year(),
  day_labels = c("S", "M", "T", "W", "T", "F", "S"),
  labels_layer = TRUE,
  color = "grey35",
  size = 3,
  alpha = 0.5
)
}
\arguments{
\item{dates_df}{}

\item{day_labels}{}

\item{labels_layer}{}

\item{color}{}

\item{size}{}

\item{alpha}{}
}
\description{
Title
}
