% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcalendar.R
\name{ggcalendar}
\alias{ggcalendar}
\title{Title}
\usage{
ggcalendar(dates_df = return_dates_year(2022))
}
\arguments{
\item{dates_df}{}
}
\value{

}
\description{
Title
}
\examples{
library(lubridate)
library(dplyr)
library(ggplot2)

ggcalendar() +
geom_text_calendar()

ggcalendar() +
geom_text_calendar(color = "grey35") +
geom_point_calendar(data = . \%>\%
filter(wday(date) == 3),
color = "red",  alpha = .4, size = 5) +
labs(title = "When to do #TidyTuesday in 2022") +
geom_text_calendar(label = "X",
                  data = data.frame(date = seq(as.Date("2022/01/01"),
                  as.Date("2022/04/18"), "days")))

}
